add move option instead of delete

+++ DupImageDeleter/DupImageDeleter/MainForm.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using System.Security.Cryptography;
 
 namespace DupImageDeleter
 {
@@ -48,8 +49,17 @@ namespace DupImageDeleter
                 this.txtExtension.Text = null;
             }
 
+            this.txtMoveDirectory.Enabled = chkMoveInsteadOfDelete.Checked;
+            this.btnMoveDirectory.Enabled = chkMoveInsteadOfDelete.Checked;
+
+            if (!this.txtMoveDirectory.Enabled)
+            {
+                this.txtMoveDirectory.Text = null;
+            }
+
             this.btnGo.Enabled = 
                 !string.IsNullOrWhiteSpace(this.txtImageDirectory.Text) 
+                && (!this.chkMoveInsteadOfDelete.Checked || !string.IsNullOrWhiteSpace(this.txtMoveDirectory.Text))
                 && 
                 (
                     this.chkDeleteFilesWithSameName.Checked && !string.IsNullOrWhiteSpace(this.txtExtension.Text)
@@ -57,9 +67,13 @@ namespace DupImageDeleter
                 );
         }
 
-        private void WalkDirectoryTree(DirectoryInfo root, string extension, bool testMode)
+        private void WalkDirectoryTree(DirectoryInfo root)
         {
-            FileInfo[] files = null;
+            string extension = this.txtExtension.Text;
+            bool testMode = chkTestMode.Checked;
+            bool move = chkMoveInsteadOfDelete.Checked;
+
+           FileInfo[] files = null;
 
             try
             {
@@ -78,6 +92,7 @@ namespace DupImageDeleter
                                           FileName = f.Name,
                                           FileNameWithoutExtension = Path.GetFileNameWithoutExtension(f.Name),
                                           FileExtension = f.Extension,
+                                          FileHash = GetMD5HashFromFile(f.FullName),
                                           FileCreationTime = f.CreationTimeUtc
                                       }).ToList();
 
@@ -95,6 +110,20 @@ namespace DupImageDeleter
                 fileAttributes.RemoveAll(x => filesToDelete.Contains(x));
                 //******************************************************************
 
+                ////******************************************************************
+                //filesToDelete.AddRange(
+                //    fileAttributes
+                //        .GroupBy(f => f.FileHash)
+                //        .Where(group => group.Count() > 1)
+                //        .Select(group =>
+                //            fileAttributes
+                //                .Where(x => string.Equals(x.FileHash, group.Key, StringComparison.OrdinalIgnoreCase))
+                //                .OrderByDescending(x => x.FileCreationTime)
+                //                .ThenByDescending(x => x.FileName).Take(1).SingleOrDefault()).ToList());
+
+                //fileAttributes.RemoveAll(x => filesToDelete.Contains(x));
+                ////******************************************************************
+
                 foreach (var fileToDelete in filesToDelete)
                 {
                     if (fileToDelete != null)
@@ -105,15 +134,26 @@ namespace DupImageDeleter
                         {
                             if (testMode)
                             {
-                                this.AddToOutput($"Would delete: {file.FullName}");
+                                string desc = move ? "move" : "delete";
+
+                                this.AddToOutput($"Would {desc}: {file.FullName}");
                             }
                             else
                             {
-                                this.AddToOutput($"Deleting: {file.FullName}");
+                                string desc = move ? "Moving" : "Deleting";
+
+                                this.AddToOutput($"{desc}: {file.FullName}");
 
                                 try
                                 {
-                                    file.Delete();
+                                    if(move)
+                                    {
+                                        file.MoveTo(this.txtMoveDirectory.Text);
+                                    }
+                                    else
+                                    {
+                                        file.Delete();
+                                    }
                                 }
                                 catch (Exception e)
                                 {
@@ -130,11 +170,23 @@ namespace DupImageDeleter
                 {
                     if(!string.Equals(dirInfo.Name, "Cache", StringComparison.OrdinalIgnoreCase))
                     {
-                        this.WalkDirectoryTree(dirInfo, extension, testMode);
+                        this.WalkDirectoryTree(dirInfo);
                     }
                 }
             }
         }
+
+        private static string GetMD5HashFromFile(string fileName)
+        {
+            using (var md5 = MD5.Create())
+            {
+                using (var stream = File.OpenRead(fileName))
+                {
+                    return BitConverter.ToString(md5.ComputeHash(stream)).Replace("-", string.Empty);
+                }
+            }
+        }
+
 
         private void AddToOutput(string output)
         {
@@ -151,23 +203,42 @@ namespace DupImageDeleter
             {
                 Cursor.Current = Cursors.WaitCursor;
 
-                WalkDirectoryTree(root, this.txtExtension.Text, chkTestMode.Checked);
+                this.WalkDirectoryTree(root);
 
                 this.AddToOutput("Finished!");
             }
             finally
             {
                 Cursor.Current = Cursors.Default;
-
             }
         }
 
-        private void txtImageDirectory_Leave(object sender, EventArgs e)
+        private void chkMoveInsteadOfDelete_CheckedChanged(object sender, EventArgs e)
         {
             this.InitControls();
         }
 
-        private void txtExtension_Leave(object sender, EventArgs e)
+        private void btnMoveDirectory_Click(object sender, EventArgs e)
+        {
+            DialogResult result = moveDirectoryBrowser.ShowDialog();
+
+            if (result == DialogResult.OK)
+            {
+                this.txtMoveDirectory.Text = moveDirectoryBrowser.SelectedPath;
+            }
+        }
+
+        private void txtImageDirectory_TextChanged(object sender, EventArgs e)
+        {
+            this.InitControls();
+        }
+
+        private void txtExtension_TextChanged(object sender, EventArgs e)
+        {
+            this.InitControls();
+        }
+
+        private void txtMoveDirectory_TextChanged(object sender, EventArgs e)
         {
             this.InitControls();
         }
